hello-8bit-world/pyproject.toml
[tool.poetry]
name = "hello-8bit-world"
version = "0.1.0"
description = ""
authors = ["chris-young-12-nhs <chris.young12@nhs.net>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pyserial = "^3.5"
python-dotenv = "^1.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
